+++ src/frontend/apps/e2e/__tests__/app-desk/header.spec.ts
@@ -11,6 +11,13 @@ test.describe('Header', () => {
   test('checks all the elements are visible', async ({ page }) => {
     const header = page.locator('header').first();
 
-    await expect(header.getByAltText('Régie Logo')).toBeVisible();
+    await expect(
+      header.getByRole('button', {
+        name: /open the menu|close the menu|ouvrir le menu|fermer le menu/i,
+      }),
+    ).toBeVisible();
+    await expect(
+      header.getByRole('button', { name: 'Les services de LaSuite' }),
+    ).toBeVisible();
   });
 });