+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to
 
 ### Changed
 
+- 💄(ui) review ui : delete border container
 - 💄(config) update ui configuration domain modale
 - 🍱(static) update logo in invitation email template #1085
 - ✨(uiV2) use Lasuite UI kit, new layout